Generate commands rather than macros by default
This allows us to inject things into the middle of commands
that we can't do with macros.  For instance, if we need to
inject a library before the LDFLAGS, like this:

  : ... |> $(CXX) $(CPPFLAGS) $(CXXFLAGS) -o %o -c %f ../libx.so $(LDFLAGS) |> ...

we can do it when generating a command, but if we're emitting
macros we can't:

  : ... |> !link ../libx.so |> ...

The macro would result in undefined externs if ../libx.so depends
on any libraries since those libraries will have been already
processed by time the linker sees libx.so.