Added run-test --disabled-by switch
This allows the switch to be used more easily with a command that
conditionally disables a test. For instance, you can say

   run-test --disabled-by=is_disabled ...

which means that test is disabled if the is_disabled command emitted
any output, but is enabled otherwise. The output of the command
becomes the reason for why the test is disabled.